Move TensorImpl to core.
