+++ aten/src/TH/THHalf.h
@@ -4,7 +4,7 @@
 #include <TH/THGeneral.h>
 
 #ifdef __cplusplus
-#include <ATen/TensorImpl.h>
+#include <ATen/core/TensorImpl.h>
 #endif
 
 #ifdef __cplusplus
